+++ pruebas/01-reading-list/SitoNumbis/src/contexts/LanguageProvider.jsx
@@ -10,6 +10,12 @@ import PropTypes from "prop-types";
 
 const LanguageContext = createContext();
 
+/**
+ *
+ * @param {object} languageState
+ * @param {object} action
+ * @returns
+ */
 const languageReducer = (languageState, action) => {
   switch (action.type) {
     case "set":
