+++ pruebas/01-reading-list/SitoNumbis/src/utils/utils.js
@@ -1,22 +1,9 @@
 import config from "../config";
 
-export const utilsToggleTheme = () => {
-  // On page load or when changing themes, best to add inline in `head` to avoid FOUC
-  if (
-    localStorage.theme === "dark" ||
-    (!("theme" in localStorage) &&
-      window.matchMedia("(prefers-color-scheme: dark)").matches)
-  ) {
-    document.documentElement.classList.add("light");
-    document.documentElement.classList.remove("dark");
-    localStorage.setItem("theme", "light");
-  } else {
-    document.documentElement.classList.add("dark");
-    document.documentElement.classList.remove("light");
-    localStorage.setItem("theme", "dark");
-  }
-};
-
+/**
+ *
+ * @param {Map<String, String>} readingList
+ */
 export const saveReadingListToLocal = (readingList) => {
   const obj = Object.fromEntries(readingList);
   const stringify = JSON.stringify(Object.keys(obj));
