+++ pruebas/01-reading-list/SitoNumbis/src/views/Home.jsx
@@ -1,13 +1,13 @@
-import { Suspense, useEffect, useMemo } from "react";
+import { Suspense, useMemo } from "react";
 import loadable from "@loadable/component";
 
 // components
+import HomeHeader from "./HomeHeader";
 import Loading from "../components/Loading/Loading";
 import FilterBar from "../components/FilterBar/FilterBar";
 
 // contexts
 import { LightBoxProvider } from "../components/LightBox/LightBoxProvider";
-import { useLanguage } from "../contexts/LanguageProvider";
 import { useLibrary } from "../contexts/LibraryProvider";
 
 // styles
@@ -19,34 +19,11 @@ const ToTop = loadable(() => import("../components/ToTop/ToTop"));
 const LightBox = loadable(() => import("../components/LightBox/LightBox"));
 
 function Home() {
-  const { libraryState, setLibraryState } = useLibrary();
-
-  const { languageState } = useLanguage();
+  const { libraryState } = useLibrary();
 
   const bookToPrints = useMemo(() => {
     return libraryState.books;
   }, [libraryState.books]);
-
-  const totalLength = useMemo(() => {
-    return libraryState.seeing === "all"
-      ? libraryState.available
-      : libraryState.readingList.size;
-  }, [
-    libraryState.seeing,
-    libraryState.available,
-    libraryState.readingList.size,
-  ]);
-
-  // get the current showing items
-  useEffect(() => {
-    // milliseconds for the dom
-    setTimeout(() => {
-      setLibraryState({
-        type: "set-showing",
-        showing: document.querySelectorAll(".book").length,
-      });
-    }, 100);
-  }, [libraryState.filtering, libraryState.seeing, setLibraryState]);
 
   return (
     <main className={styles.main}>
@@ -54,23 +31,7 @@ function Home() {
         fallback={<Loading className="w-full h-full top-0 left-0 fixed z-50" />}
       >
         <FilterBar />
-        <div className="flex items-center flex-wrap w-full">
-          {/* current list total */}
-          <p>
-            {languageState.texts.seeing.title}{" "}
-            {languageState.texts.seeing[libraryState.seeing]}
-            <span className="alter-text text-sm mx-2">({totalLength})</span>
-          </p>
-          {/* If the user is using the genre filter */}
-          {libraryState.filtering.length ? (
-            <p>
-              {">"} {libraryState.filtering}{" "}
-              <span className="alter-text text-sm mx-1">
-                ({libraryState.showing})
-              </span>
-            </p>
-          ) : null}
-        </div>
+        <HomeHeader />
         <LightBoxProvider>
           <LightBox />
           <section>
